+++ src/tests/scatteraction.cc
@@ -84,7 +84,8 @@ TEST(elastic_collision) {
   constexpr NNbarTreatment nnbar_treatment = NNbarTreatment::NoAnnihilation;
   act.add_all_scatterings(sigma, true, Test::all_reactions_included(),
                           Test::no_multiparticle_reactions(), 0.,
-                          strings_switch, false, false, nnbar_treatment);
+                          strings_switch, false, false, nnbar_treatment, 1.0,
+                          0.0);
 
   // check cross section
   COMPARE(act.cross_section(), sigma);
@@ -157,10 +158,10 @@ TEST(outgoing_valid) {
   constexpr double elastic_parameter = 0.;  // don't include elastic scattering
   constexpr bool strings_switch = false;
   constexpr NNbarTreatment nnbar_treatment = NNbarTreatment::NoAnnihilation;
-  act->add_all_scatterings(elastic_parameter, true,
-                           Test::all_reactions_included(),
-                           Test::no_multiparticle_reactions(), 0.,
-                           strings_switch, false, false, nnbar_treatment);
+  act->add_all_scatterings(
+      elastic_parameter, true, Test::all_reactions_included(),
+      Test::no_multiparticle_reactions(), 0., strings_switch, false, false,
+      nnbar_treatment, 1.0, 0.0);
 
   VERIFY(act->cross_section() > 0.);
 
@@ -234,11 +235,11 @@ TEST(cross_sections_symmetric) {
     act12->add_all_scatterings(elastic_parameter, two_to_one, included_2to2,
                                included_multi, low_snn_cut, strings_switch,
                                use_AQM, strings_with_probability,
-                               nnbar_treatment);
+                               nnbar_treatment, 1.0, 0.0);
     act21->add_all_scatterings(elastic_parameter, two_to_one, included_2to2,
                                included_multi, low_snn_cut, strings_switch,
                                use_AQM, strings_with_probability,
-                               nnbar_treatment);
+                               nnbar_treatment, 1.0, 0.0);
 
     VERIFY(act12->cross_section() >= 0.);
     VERIFY(act21->cross_section() >= 0.);
@@ -294,7 +295,8 @@ TEST(pythia_running) {
   constexpr NNbarTreatment nnbar_treatment = NNbarTreatment::NoAnnihilation;
   act->add_all_scatterings(elastic_parameter, false, ReactionsBitSet(),
                            Test::no_multiparticle_reactions(), 0.,
-                           strings_switch, false, false, nnbar_treatment);
+                           strings_switch, false, false, nnbar_treatment, 1.0,
+                           0.0);
 
   VERIFY(act->cross_section() > 0.);
 
@@ -364,10 +366,10 @@ TEST(no_strings) {
         0.;  // don't include elastic scattering
     constexpr bool strings_switch = false;
     constexpr NNbarTreatment nnbar_treatment = NNbarTreatment::NoAnnihilation;
-    act->add_all_scatterings(elastic_parameter, true,
-                             Test::all_reactions_included(),
-                             Test::no_multiparticle_reactions(), 0.,
-                             strings_switch, false, false, nnbar_treatment);
+    act->add_all_scatterings(
+        elastic_parameter, true, Test::all_reactions_included(),
+        Test::no_multiparticle_reactions(), 0., strings_switch, false, false,
+        nnbar_treatment, 1.0, 0.0);
 
     VERIFY(act->cross_section() > 0.);
 
@@ -410,7 +412,7 @@ TEST(update_incoming) {
   NNbarTreatment nnbar_treatment = NNbarTreatment::NoAnnihilation;
   act.add_all_scatterings(sigma, true, Test::all_reactions_included(),
                           Test::no_multiparticle_reactions(), 0., string_switch,
-                          false, false, nnbar_treatment);
+                          false, false, nnbar_treatment, 1.0, 0.0);
 
   // change the position of one of the particles
   const FourVector new_position(0.1, 0., 0., 0.);
@@ -489,11 +491,11 @@ TEST(particle_ordering) {
     act12->add_all_scatterings(elastic_parameter, two_to_one, included_2to2,
                                included_multi, low_snn_cut, strings_switch,
                                use_AQM, strings_with_probability,
-                               nnbar_treatment);
+                               nnbar_treatment, 1.0, 0.0);
     act21->add_all_scatterings(elastic_parameter, two_to_one, included_2to2,
                                included_multi, low_snn_cut, strings_switch,
                                use_AQM, strings_with_probability,
-                               nnbar_treatment);
+                               nnbar_treatment, 1.0, 0.0);
 
     VERIFY(act12->cross_section() >= 0.);
     VERIFY(act21->cross_section() >= 0.);